Conflicts:
	src/main/java/de/unistuttgart/vis/vita/analysis/modules/ANNIEModule.java
	src/main/java/de/unistuttgart/vis/vita/analysis/modules/EntityRelationModule.java
	src/main/java/de/unistuttgart/vis/vita/analysis/modules/EntityWordCloudModule.java
	src/main/java/de/unistuttgart/vis/vita/model/document/Chapter.java
	src/main/java/de/unistuttgart/vis/vita/model/document/TextSpan.java
	src/main/java/de/unistuttgart/vis/vita/model/entity/Attribute.java
	src/main/java/de/unistuttgart/vis/vita/model/entity/BasicEntity.java
	src/main/java/de/unistuttgart/vis/vita/services/occurrence/AttributeOccurrencesService.java
	src/main/java/de/unistuttgart/vis/vita/services/occurrence/EntityOccurrencesService.java
	src/main/java/de/unistuttgart/vis/vita/services/occurrence/RelationOccurrencesService.java
	src/test/java/de/unistuttgart/vis/vita/model/TextSpanTest.java

+++ src/test/java/de/unistuttgart/vis/vita/analysis/importer/TextImportModuleTest.java
@@ -74,7 +74,7 @@ public class TextImportModuleTest {
     DocumentMetadata metadata = result.getMetadata();
     assertEquals("Matilde Serao", metadata.getAuthor());
     assertEquals("The conquest of Rome", metadata.getTitle());
-    assertEquals(2014, metadata.getPublishYear());
+    assertEquals(2014, metadata.getPublishYear().intValue());
     assertEquals("", metadata.getEdition());
     assertEquals("", metadata.getGenre());
     assertEquals("", metadata.getPublisher());
@@ -117,7 +117,7 @@ public class TextImportModuleTest {
     DocumentMetadata metadata = result.getMetadata();
     assertEquals("John Buchan", metadata.getAuthor());
     assertEquals("The Thirty-nine Steps", metadata.getTitle());
-    assertEquals(1996, metadata.getPublishYear());
+    assertEquals(1996, metadata.getPublishYear().intValue());
     assertEquals("", metadata.getEdition());
     assertEquals("", metadata.getGenre());
     assertEquals("", metadata.getPublisher());
@@ -146,7 +146,7 @@ public class TextImportModuleTest {
     DocumentMetadata metadata = result.getMetadata();
     assertEquals("Matilde Serao", metadata.getAuthor());
     assertEquals("The conquest of Rome", metadata.getTitle());
-    assertEquals(2014, metadata.getPublishYear());
+    assertEquals(2014, metadata.getPublishYear().intValue());
     assertEquals("", metadata.getEdition());
     assertEquals("", metadata.getGenre());
     assertEquals("", metadata.getPublisher());
